Assistant checkpoint
Assistant generated file changes:
- client/src/pages/Home.tsx: Fix date comparison logic in isArchived function

---

User prompt:

i'm not sure if we need to do that, but i note that it works, whereas it's still the case that new events being created with future dates are still only showing up in the archive. we've tried at least 10 times to fix this. there must be some underlying issue that we introduced since last deployment, and we need to track it down